pyproject.toml
[project]
name = "django-in-product"
version = "0.1.0"
description = "Django in Product"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "django==5.2.8",
    "djangorestframework==3.16.1",
]

[dependency-groups]
dev = [
    "django-stubs==5.2.7",
    "djangorestframework-stubs==3.16.5",
    "mypy==1.18.2",
    "pytest==9.0.1",
    "pytest-django==4.11.1",
    "ruff==0.14.6",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests/*.py"]
pythonpath = ["backend"]

[tool.ruff.lint]
ignore = [
]

select = [
   "I",
]
